+++ NEWS.md
@@ -17,7 +17,7 @@
   lowercased value is used; if inferred, the last element of the SVN project
   name (after splitting by -) is used (lowercased). The project parameter within
   the report is always set to the last part of the SVN project URL, regardless
-  of .project_number. Example: one-qc-report-YYYY-MM-DD.pdf. (#999)
+  of .project_number. Example: one-qc-report-YYYY-MM-DD.pdf. (#162)
 
 # review 3.10.0